pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Amirali Lotfi"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-google-genai (>=2.1.8,<3.0.0)",
    "langchain-core (>=0.3.0,<0.4.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langgraph (>=0.2.60,<0.3.0)",
    "openai (>=1.40.0,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
langsmith = "^0.4.6"
python-docx = "^1.2.0"
reportlab = "^4.4.3"
arabic-reshaper = "^3.0.0"
python-bidi = "^0.6.6"
xlsxwriter = "^3.2.5"
openpyxl = "^3.1.5"

[tool.poetry]
package-mode = false
